v0.4.89-experimental ⬇
just a draft (for fun basically) on how it can be the hypothetical server configuration file.
it is named _.koorierc for developing purpose, renamed it to .koorierc, adapt the content of the file at your needs and run npx koorie ❗without any flags

- example of .koorierc file:

; this is a comment nad will be ignored form being parsed

static_files = path/to/public
; will fail because the option (left operand) must be the same, without double hyphen prefixing, as the flag passed to the command line [--static-files]. ❗ ONLY use the long flags!

s = path/to/public
; will fail ❗ DO NOT USE short flags!

static-files = path/to/public
; will go through ;)

static-files= path/to/public
; will fail because ❗ MUST be space from the equal sign for both left operand and right operand!

- example ends

just a note:

The parser is just for fun nothing serious, but it looks for charCode = 32 at position 0 of the next line read from node:readline interface, that corresponds to one empty character( one space ).
In that case everything after the first space will be treated as a comment so then skipped form being parsed.

big changes, ❗ not safe for production!
test test test
profiling profiling profiling